Added the functionality to refresh token.
Upon login, we are adding auth token to local storage now expires at is also added. Also added a new property to global this "lastInteracted". This property is updated when any runevent is executed. Lastly in index added an interval function which runs every 60 seconds to see if the token is about to expire in 3 minutes and there is some interaction happened in last 15 minutes. If both the criteria met then the token is refreshed and set in local storage for other tabs to pickup the new token and use.